+++ src/app/pencil_init.py
@@ -1,0 +1,6 @@
+
+class PencilInit:
+    class CommonUser:
+        def __init__(self, app):
+            self.app = app
+            return
